+++ mobilityteamproject/Step3_Raspberry_Pi_Porting/labels.txt
@@ -1,0 +1,2 @@
+0 wearing a helmet
+1 not wearing a helmet
