+++ mobilityteamproject/Step3_Raspberry_Pi_Porting/requirements.txt
@@ -1,0 +1,15 @@
+edgetpu @ file:///home/pi/Downloads/edgetpu-2.12.1-py3-none-any.whl
+
+imutils==0.5.3
+
+numpy==1.19.1
+
+opencv-contrib-python==4.1.0.25
+
+picamera==1.13
+
+Pillow==7.2.0
+
+simpleaudio==1.0.4
+
+tflite-runtime @ https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp37-cp37m-linux_armv7l.whl
